pyproject.toml
[project]
name = "ai-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nicegui>=2.22.2",
    "openai>=1.99.9",
    "pinecone>=7.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pywebview>=6",
]
